enable ssl connections through IRC-js api
this allows the user to include an `ssl` hash (or bool) in the server
connection params:

```js
var irc = require('irc-js');
var bot = irc.connect({
  server: {
    address: 'chat.freenode.net',
    port: 6697,
    ssl: true // could also be: {rejectUnauthorized: false}
  },
  nick: 'aaronj1335'
}, function(bot) {
  // ...
});
```

the value defaults to false, and if it's truthy, it gets passed strait
through to [`tls.connect`][tls-connect].

a couple thoughts:

- i didn't include a unit test only because i don't know what exactly to
test, though i'm happy to implement something

- this messes with the server caching. cases where the user wants two
connections, one ssl and one not ssl to the same server/port will be
handled incorrectly. i think these are rare enough that it's safe to
ignore for now though.

[tls-connect]: http://nodejs.org/api/tls.html